remove getDataForSearchInput as utils method and move it as method of SearchPanel.vue
